+++ changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mock types with explicit and non-explicit properties/indexers now generate correct code (issue [#299](https://github.com/JasonBock/Rocks/issues/299))
 - Ignored a warning when a make was generated and `[MemberNotNullWhen]` existed (issue [#297](https://github.com/JasonBock/Rocks/issues/297))
 - Added generics to event extension methods when needed (issue [#295](https://github.com/JasonBock/Rocks/issues/295))
+- Fixed a variable naming issue in the member expectation handlers (issue [#296](https://github.com/JasonBock/Rocks/issues/296))
 
 ## [8.0.0-alpha.2] - 2024-01-14